+++ jpkc/src/com/jpkc/web/controller/TeamController.java
@@ -10,6 +10,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.jpkc.commons.Page;
 import com.jpkc.commons.Render;
@@ -21,13 +23,12 @@ import com.jpkc.util.VerifyUtil;
 @Controller
 @RequestMapping("/team")
 public class TeamController {
-	
+
 	private static Log log = LogFactory.getLog(TeamController.class);
-	
+
 	@Resource
 	private TeamService teamService;
-	
-	
+
 	/**
 	 * 团队成员列表
 	 * 
@@ -35,34 +36,34 @@ public class TeamController {
 	 * @2015-11-8
 	 */
 	@RequestMapping("/list")
-	public String teamList(Model model,String memberName) {
+	public String teamList(Model model, String memberName) {
 		Team team = new Team();
-		if(!VerifyUtil.isEmpty(memberName)) {
+		if (!VerifyUtil.isEmpty(memberName)) {
 			team.setMemberName(memberName);
 		}
 		int pageSize = 10;
-		int pageNumber= 1;
-		Page<Team> page = teamService.getTeamByPage(team,pageSize,pageNumber);
+		int pageNumber = 1;
+		Page<Team> page = teamService.getTeamByPage(team, pageSize, pageNumber);
 		model.addAttribute("page", page);
 		model.addAttribute("memberName", memberName);
-		
+
 		return "/console/team/teamlist";
 	}
-	
+
 	/**
-	 * 团队成员列表
+	 * 团队成员列表分页
 	 * 
 	 * @author zhangyi
 	 * @2015-11-8
 	 */
 	@RequestMapping("/list/{pageNumber}/{pageSize}")
-	public String teamList(Model model,@PathVariable int pageNumber, @PathVariable int pageSize ) {
+	public String teamList(Model model, @PathVariable int pageNumber, @PathVariable int pageSize) {
 		Team team = new Team();
-		Page<Team> page = teamService.getTeamByPage(team,pageSize,pageNumber);
+		Page<Team> page = teamService.getTeamByPage(team, pageSize, pageNumber);
 		model.addAttribute("page", page);
 		return "console/team/teamlist";
 	}
-	
+
 	/**
 	 * 团队成员列表API
 	 * 
@@ -70,16 +71,16 @@ public class TeamController {
 	 * @2015-11-8
 	 */
 	@RequestMapping("/teamList")
-	public Render<Object> teamList(Model model,int pageNumber,int pageSize,int type) {
+	public Render<Object> teamList(Model model, int pageNumber, int pageSize, int type) {
 		Render<Object> render = new Render<Object>();
 		Team team = new Team();
 		team.setType(type);
-		Page<Team> page = teamService.getTeamByPage(team,pageSize,pageNumber);
+		Page<Team> page = teamService.getTeamByPage(team, pageSize, pageNumber);
 		render.setData(page);
 		render.setCode("20000");
 		return render;
 	}
-	
+
 	/**
 	 * 添加
 	 * 
@@ -88,28 +89,36 @@ public class TeamController {
 	 */
 	@RequestMapping("/add")
 	public String addIndex() {
-		
+
 		return "/console/team/add";
 	}
-	
+
 	/**
 	 * 添加成员
 	 * 
 	 * @author zhangyi
 	 * @2015-11-13
 	 */
-	@RequestMapping("/addTeam")
-	public String addTeam(Model model,Team team) {
+	@RequestMapping(value="/addTeam",method=RequestMethod.POST)
+	public String addTeam(Model model, Team team) {
 		User user = new User();
 		user.setUsername("123");
 		team.setCreateBy(user.getUsername());
 		teamService.add(team);
-		
+
 		return "redirect:/team/list";
 	}
 	
+	/**
+	 * 
+	 * 编辑
+	 * 
+	 * @param id
+	 * @param model
+	 * @return
+	 */
 	@RequestMapping("/edit/{id}")
-	public String editTeam(@PathVariable int id,Model model) {
+	public String editTeam(@PathVariable int id, Model model) {
 		Team team = new Team();
 		team.setId(id);
 		List<Team> teamList = teamService.select(team);
@@ -117,11 +126,30 @@ public class TeamController {
 		return "console/team/add";
 	}
 	
-	@RequestMapping("/del/{id}")
-	public String delNotice(@PathVariable int id) {
-		Team team = new Team();
-		team.setId(id);
-		teamService.delete(team);
+	/**
+	 * 
+	 * 批量删除
+	 * 
+	 * @param ids
+	 * @param ra
+	 * @return
+	 */
+	@RequestMapping("/del/{ids}")
+	public String delNotice(@PathVariable int[] ids, RedirectAttributes ra) {
+		Render<Object> render = new Render<Object>();
+		if (VerifyUtil.isEmpty(ids)) {
+			log.info("删除人员信息的id是空!");
+			render.setCode("20000");
+			render.setData("删除人员信息id是空!");
+			ra.addFlashAttribute("render", render);
+			return "redirect:/team/list";
+		}
+
+		for (int i = 0; i < ids.length; i++) {
+			Team team = new Team(ids[i]);
+			teamService.delete(team);
+		}
+
 		return "redirect:/team/list";
 	}
 }
